+++ aas-sm-to-py/aas_sm_to_py/code_templates/referable_with_se_class.pyi
