+++ aas-sm-to-py/aas_sm_to_py/code_templates/se_col_class.pyi
