+++ aas-sm-to-py/aas_sm_to_py/code_templates/se_list_class.pyi
