+++ aas-sm-to-py/aas_sm_to_py/submodel_to_code.py
@@ -3,7 +3,7 @@
 import argparse
 import pathlib
 
-from asttp import generator
+from aas_sm_to_py import generator
 
 
 def main() -> None:
