+++ aas-sm-to-py/aas_sm_to_py/util.py
