+++ py-aas-submodels/py_aas_submodels/control_component_type_02015-1-0.py
