+++ py-aas-submodels/py_aas_submodels/nameplate_02006-2-0.py
